server logic refactoring, minor adjustments

+++ server/src/models/tripModel.ts
@@ -1,16 +1,21 @@
 import db from '../db/sqlite';
-import {Trip, User} from '@shared/index';
+import { Trip } from '@shared/index';
+import { isValidId } from '../utils/dataValidation';
 
 export const getAllTrips = (): Promise<Trip[]> => {
   return new Promise((resolve, reject) => {
-    db.all('SELECT * FROM trip', [], (err, rows) => {
+    db.all('SELECT * FROM trip', [], (err, rows: Trip[]) => {
       if (err) return reject(err);
       resolve(rows.map(row => row as Trip));
     });
   });
 };
 
-export const getTripById = (id: string): Promise<Trip> => {
+export const getTripById = (id: number): Promise<Trip> => {
+  if (!isValidId(id)) {
+    return Promise.reject(new Error('Invalid ID'));
+  }
+
   return new Promise((resolve, reject) => {
     db.get('SELECT * FROM trip WHERE id = ?', [id], (err, row: Trip | undefined) => {
       if (err) return reject(err);
@@ -20,7 +25,7 @@ export const getTripById = (id: string): Promise<Trip> => {
   });
 };
 
-export const addTrip = (trip:Omit<Trip, "id"> ): Promise<number> => {
+export const addTrip = (trip:Omit<Trip, 'id'> ): Promise<number> => {
   return new Promise((resolve, reject) => {
     db.run(
         'INSERT INTO trip (name, year, location, image) VALUES (?, ?, ?, ?)',
@@ -36,6 +41,10 @@ export const addTrip = (trip:Omit<Trip, "id"> ): Promise<number> => {
 };
 
 export const deleteTrip = (tripId:number): Promise<number> => {
+  if (!isValidId(tripId)) {
+    return Promise.reject(new Error('Invalid ID'));
+  }
+
   return new Promise((resolve, reject) => {
     db.run(
         'delete from trip where id = ?',
@@ -49,7 +58,7 @@ export const deleteTrip = (tripId:number): Promise<number> => {
         });
   });
 }
-export const updateTrip = (trip:Trip ): Promise<number> => {
+export const updateTrip = (trip:Trip): Promise<number> => {
   return new Promise((resolve, reject) => {
     const { id, name, year, location} = trip;
     db.run(
